Refactor BoreholeGeometryController
