+++ tests/Controllers/BoreholeGeometryControllerTest.cs
@@ -211,113 +211,4 @@ public class BoreholeGeometryControllerTest
         var geometries = context.BoreholeGeometry.Where(g => g.BoreholeId == boreholeIdWithGeometry).OrderBy(g => g.Id).ToList();
         Assert.AreEqual(0, geometries.Count());
     }
-
-    [TestMethod]
-    public void ConvertAzIncToXYZ()
-    {
-        var azimuthInclinationData = new List<BoreholeGeometryController.AzInc.Geometry>
-        {
-            new BoreholeGeometryController.AzInc.Geometry
-            {
-                MeasuredDepth = 3500,
-                Azimuth = ToRadians(20),
-                Inclination = ToRadians(15),
-            },
-            new BoreholeGeometryController.AzInc.Geometry
-            {
-                MeasuredDepth = 3600,
-                Azimuth = ToRadians(45),
-                Inclination = ToRadians(25),
-            },
-        };
-
-        var xyzData = BoreholeGeometryController.AzInc.ConvertToXYZ(azimuthInclinationData);
-
-        Assert.AreEqual(2, xyzData.Count);
-        Assert.AreEqual(0, xyzData[0].X);
-        Assert.AreEqual(0, xyzData[0].Y);
-        Assert.AreEqual(0, xyzData[0].Z);
-        Assert.AreEqual(19.451, xyzData[1].X, 0.01);
-        Assert.AreEqual(27.218, xyzData[1].Y, 0.01);
-        Assert.AreEqual(94.012, xyzData[1].Z, 0.01);
-    }
-
-    [TestMethod]
-    public void ConvertAzIncToXYZDuplicateEntry()
-    {
-        var azimuthInclinationData = new List<BoreholeGeometryController.AzInc.Geometry>
-        {
-            new BoreholeGeometryController.AzInc.Geometry
-            {
-                MeasuredDepth = 3500,
-                Azimuth = ToRadians(20),
-                Inclination = ToRadians(15),
-            },
-            new BoreholeGeometryController.AzInc.Geometry
-            {
-                MeasuredDepth = 3500,
-                Azimuth = ToRadians(20),
-                Inclination = ToRadians(15),
-            },
-        };
-
-        var xyzData = BoreholeGeometryController.AzInc.ConvertToXYZ(azimuthInclinationData);
-
-        Assert.AreEqual(2, xyzData.Count);
-        Assert.AreEqual(0, xyzData[0].X);
-        Assert.AreEqual(0, xyzData[0].Y);
-        Assert.AreEqual(0, xyzData[0].Z);
-        Assert.AreEqual(0, xyzData[1].X, 0.01);
-        Assert.AreEqual(0, xyzData[1].Y, 0.01);
-        Assert.AreEqual(0, xyzData[1].Z, 0.01);
-    }
-
-    [TestMethod]
-    public void ConvertRollPitchYawToAzIncStraightDown()
-    {
-        var data = new List<BoreholeGeometryController.PitchRoll.Geometry>
-        {
-            new BoreholeGeometryController.PitchRoll.Geometry
-            {
-                MeasuredDepth = 1200,
-                Roll = ToRadians(0),
-                Pitch = ToRadians(0),
-                MagneticRotation = ToRadians(0),
-            },
-        };
-
-        var azIncData = BoreholeGeometryController.PitchRoll.ConvertToAzInc(data);
-
-        Assert.AreEqual(1, azIncData.Count);
-        Assert.AreEqual(1200, azIncData[0].MeasuredDepth);
-        Assert.AreEqual(0, azIncData[0].Azimuth);
-        Assert.AreEqual(0, azIncData[0].Inclination);
-    }
-
-    [TestMethod]
-    public void ConvertRollPitchYawToAzInc()
-    {
-        var data = new List<BoreholeGeometryController.PitchRoll.Geometry>
-        {
-            new BoreholeGeometryController.PitchRoll.Geometry
-            {
-                MeasuredDepth = 0,
-                Roll = ToRadians(-6.3),
-                Pitch = ToRadians(8.5),
-                MagneticRotation = ToRadians(99.4),
-            },
-        };
-
-        var azIncData = BoreholeGeometryController.PitchRoll.ConvertToAzInc(data);
-
-        Assert.AreEqual(1, azIncData.Count);
-        Assert.AreEqual(0, azIncData[0].MeasuredDepth);
-        Assert.AreEqual(ToRadians(136.16), azIncData[0].Azimuth, 1e-4);
-        Assert.AreEqual(ToRadians(10.57), azIncData[0].Inclination, 1e-4);
-    }
-
-    private double ToRadians(double degrees)
-    {
-        return degrees * Math.PI / 180;
-    }
 }
